Change formatting of braces in source files.

+++ src/main/java/net/coobird/thumbnailator/util/exif/ExifUtils.java
@@ -18,8 +18,7 @@ import org.w3c.dom.NodeList;
  * @author coobird
  *
  */
-public final class ExifUtils
-{
+public final class ExifUtils {
 	
 	private static final String EXIF_MAGIC_STRING = "Exif";
 	
@@ -42,55 +41,46 @@ public final class ExifUtils
 	 * 									have the target image set, or if the
 	 * 									reader does not have a JPEG open.
 	 */
-	public static Orientation getExifOrientation(ImageReader reader, int imageIndex) throws IOException
-	{
+	public static Orientation getExifOrientation(ImageReader reader, int imageIndex) throws IOException {
 		IIOMetadata metadata = reader.getImageMetadata(imageIndex);
 		Node rootNode = metadata.getAsTree("javax_imageio_jpeg_image_1.0");
-		
+
 		NodeList childNodes = rootNode.getChildNodes();
-		
+
 		// Look for the APP1 containing Exif data, and retrieve it.
-		for (int i = 0; i < childNodes.getLength(); i++)
-		{
-			if ("markerSequence".equals(childNodes.item(i).getNodeName()))
-			{
+		for (int i = 0; i < childNodes.getLength(); i++) {
+			if ("markerSequence".equals(childNodes.item(i).getNodeName())) {
 				NodeList markerSequenceChildren = childNodes.item(i).getChildNodes();
-				
-				for (int j = 0; j < markerSequenceChildren.getLength(); j++)
-				{
-					IIOMetadataNode metadataNode = (IIOMetadataNode)(markerSequenceChildren.item(j));
-					
-					byte[] bytes = (byte[])metadataNode.getUserObject();
-					if (bytes == null)
-					{
+
+				for (int j = 0; j < markerSequenceChildren.getLength(); j++) {
+					IIOMetadataNode metadataNode = (IIOMetadataNode) (markerSequenceChildren.item(j));
+
+					byte[] bytes = (byte[]) metadataNode.getUserObject();
+					if (bytes == null) {
 						continue;
 					}
-					
+
 					byte[] magicNumber = new byte[4];
 					ByteBuffer.wrap(bytes).get(magicNumber);
-					
-					if (EXIF_MAGIC_STRING.equals(new String(magicNumber)))
-					{
+
+					if (EXIF_MAGIC_STRING.equals(new String(magicNumber))) {
 						return getOrientationFromExif(bytes);
 					}
 				}
 			}
 		}
-		
+
 		return null;
 	}
-	
-	private static Orientation getOrientationFromExif(byte[] exifData)
-	{
-		
+
+	private static Orientation getOrientationFromExif(byte[] exifData) {
 		// Needed to make byte-wise reading easier.
 		ByteBuffer buffer = ByteBuffer.wrap(exifData);
-		
+
 		byte[] exifId = new byte[4];
 		buffer.get(exifId);
-		
-		if (!EXIF_MAGIC_STRING.equals(new String(exifId)))
-		{
+
+		if (!EXIF_MAGIC_STRING.equals(new String(exifId))) {
 			return null;
 		}
 
@@ -101,51 +91,45 @@ public final class ExifUtils
 
 		byte[] tiffHeader = new byte[8];
 		buffer.get(tiffHeader);
-		
+
 		/*
 		 * The first 2 bytes of the TIFF header contains either:
 		 *   "II" for Intel byte alignment (little endian), or
 		 *   "MM" for Motorola byte alignment (big endian)
 		 */
 		ByteOrder bo;
-		if (tiffHeader[0] == 'I' && tiffHeader[1] == 'I')
-		{
+		if (tiffHeader[0] == 'I' && tiffHeader[1] == 'I') {
 			bo = ByteOrder.LITTLE_ENDIAN;
-		}
-		else
-		{
+		} else {
 			bo = ByteOrder.BIG_ENDIAN;
 		}
-		
+
 		byte[] numFields = new byte[2];
 		buffer.get(numFields);
-		
+
 		int nFields = ByteBuffer.wrap(numFields).order(bo).getShort();
-		
+
 		byte[] ifd = new byte[12];
-		for (int i = 0; i < nFields; i++)
-		{
+		for (int i = 0; i < nFields; i++) {
 			buffer.get(ifd);
 			IfdStructure ifdStructure = readIFD(ifd, bo);
-			
+
 			// Return the orientation from the orientation IFD
-			if (ifdStructure.getTag() == 0x0112)
-			{
+			if (ifdStructure.getTag() == 0x0112) {
 				return Orientation.typeOf(ifdStructure.getOffsetValue());
 			}
 		}
-		
+
 		return null;
 	}
 
-	private static IfdStructure readIFD(byte[] ifd, ByteOrder bo)
-	{
+	private static IfdStructure readIFD(byte[] ifd, ByteOrder bo) {
 		ByteBuffer buffer = ByteBuffer.wrap(ifd).order(bo);
-		
+
 		short tag = buffer.getShort();
 		short type = buffer.getShort();
 		int count = buffer.getInt();
-		
+
 		IfdType ifdType = IfdType.typeOf(type);
 		int offsetValue = 0;
 		
@@ -154,29 +138,20 @@ public final class ExifUtils
 		 * 4 bytes, it will exist in the earlier byte.
 		 */
 		int byteSize = count * ifdType.size();
-		
+
 		if (byteSize <= 4) {
-			if (ifdType == IfdType.SHORT)
-			{
-				for (int i = 0; i < count; i++)
-				{
-					offsetValue = (int)buffer.getShort();
+			if (ifdType == IfdType.SHORT) {
+				for (int i = 0; i < count; i++) {
+					offsetValue = buffer.getShort();
 				}
-			}
-			else if (ifdType == IfdType.BYTE || ifdType == IfdType.ASCII || ifdType == IfdType.UNDEFINED)
-			{
-				for (int i = 0; i < count; i++)
-				{
-					offsetValue = (int)buffer.get();
+			} else if (ifdType == IfdType.BYTE || ifdType == IfdType.ASCII || ifdType == IfdType.UNDEFINED) {
+				for (int i = 0; i < count; i++) {
+					offsetValue = buffer.get();
 				}
-			}
-			else
-			{
+			} else {
 				offsetValue = buffer.getInt();
 			}
-		}
-		else
-		{
+		} else {
 			offsetValue = buffer.getInt();
 		}
